Split Pipeline.swft into config & processing
